+++ data-raw/randomTreeDistances.R
@@ -12,29 +12,26 @@ RandomDistances <- function (nLeaves, repls) {
                           tr2 <- RandomTree(n)
                           cat('.')
                           sprDist <- phangorn::SPR.dist(tr1, tr2)
-
-                          c(VariationOfPhylogeneticInfo(tr1, tr2, normalize=TRUE),
-                            VariationOfMatchingSplitInfo(tr1, tr2, normalize=TRUE),
-                            VariationOfClusteringInfo(tr1, tr2, normalize=TRUE),
-                            Quartet::QuartetDivergence(Quartet::QuartetStatus(tr1, tr2), similarity = FALSE),
-                            1 - NyeTreeSimilarity(tr1, tr2, normalize=TRUE),
-                            MatchingSplitDistance(tr1, tr2, normalize=FALSE),
-                            phangorn::treedist(tr1, tr2) / c(n + n - 6L, 1), # No norm for path
-                            sprDist / (n - 3),
-                            sprDist / ((n - 2) / 2)
-                            # Upper and lower bound for SPR diameter
-                            # (Allen & Steel 2001)
+                          dists <- TreeDistData:::AllDists(tr1, tr2)
+                          dists[c(1:9, 9)] /
+                            c(rep(1L, 6L), rf = n + n - 6L,
+                              path = 1L,
+                              sprUpper = n - 3L, sprLower = (n - 2L) / 2L
+                              # Upper and lower bound for SPR diameter
+                              # (Allen & Steel 2001)
                             )
                         },
                         double(10L))
-    cbind(mean = rowMeans(distances), sd = apply(distances, 1, sd),
-          min = apply(distances, 1, min),
-          max = apply(distances, 1, max))
-    }, matrix(0, ncol=4L, nrow=10L)
+    t(rbind(apply(distances, 1L, summary),
+          apply(distances, 1L, quantile, probs=c(0.01, 0.05, 0.95, 0.99)),
+          sd = apply(distances, 1L, sd)
+    )[c(1, 7, 8, 2, 3, 4, 9, 10, 6, 5, 11), ])
+    }, matrix(0, ncol=11L, nrow=10L)
   )
   dimnames(ret) <- list(c('vpi', 'vmsi', 'vci', 'qd', 'nts', 'msd', 'rf',
                           'path', 'spr', 'sprLB'),
-                        c('mean', 'sd', 'min', 'max'),
+                        c('min', '1%', '5%', '25%', '50%', '75%', '95%', '99%',
+                          'max', 'mean', 'sd'),
                         nLeaves)
   ret
 }